file structure, refactor inputs
